+++ lib/mgmg/utils.rb
@@ -47,6 +47,10 @@ module Mgmg
 		attr_accessor :equip
 	end
 	
+	module_function def exp(smith, comp)
+		((smith-1)**2) + (2*((comp-1)**2)) + 3
+	end
+	
 	CharacterList = /[^\(\)\+0123456789\[\]あきくしすたてなねのびりるイウガクグサジスタダチツデトドニノフブペボムラリルロンヴー一万二光兜典刀剣劣匠双古名吹咆品哮地大天太子安宝小帽弓弩当息悪戦手指斧書服木本杖業樹歴殺水氷法火炎牙物玉王産用界異的皮盾短石砕竜紫綿耳聖脛腕腿般良色衣袋覇質軍軽輝輪重量金鉄鎧闇陽靴額飾首骨鬼龍]/
 	module_function def check_string(str)
 		str = str.gsub(/[\s　\\]/, '')